+++ packages/fiori-app-sub-generator/src/fiori-app-generator/fioriAppGenerator.ts
@@ -178,7 +178,8 @@ export class FioriAppGenerator extends Generator {
                 let serviceAnswers = await promptOdataServiceAnswers(
                     options,
                     FioriAppGenerator.logger as Logger,
-                    this.env.adapter as unknown as Adapter
+                    this.env.adapter as unknown as Adapter,
+                    cachedService?.connectedSystem
                 );
                 /** Back button issue temp fix */
                 // Persist derived state to facilitate backwards navigation
@@ -408,6 +409,7 @@ export class FioriAppGenerator extends Generator {
                     enableCodeAssist: this.state.project?.enableCodeAssist ?? false,
                     // Assumption that npm workspaces will be enabled if cds ui5 plugin is a depenedency
                     useNpmWorkspaces: !!(
+                        this.state.project.enableTypeScript || // If typescript is enabled, it is required that the CAP project will be updated to use NPM workspaces
                         this.state.service.capService?.cdsUi5PluginInfo?.isCdsUi5PluginEnabled ||
                         this.state.service.capService?.cdsUi5PluginInfo?.hasCdsUi5Plugin ||
                         this.state.service.capService?.cdsUi5PluginInfo?.isWorkspaceEnabled