+++ packages/preview-middleware-client/test/__mock__/sap/ui/rta/util/isReuseComponent.ts
@@ -1,3 +1,1 @@
-export default {
-    isReuseComponent: jest.fn()
-}
+export default jest.fn();